requirements.txt
Flask
mysql-connector-python
Werkzeug
reportlab
python-dotenv
gunicorn